F12d - Cookies/test/no/hvl/dat108/CookieUtilTest.java
package no.hvl.dat108;

import static org.junit.jupiter.api.Assertions.assertEquals;

import javax.servlet.http.Cookie;

import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.springframework.mock.web.MockHttpServletRequest;
import org.springframework.mock.web.MockHttpServletResponse;

class CookieUtilTest {

    private MockHttpServletRequest request;
    private MockHttpServletResponse response;
    
    @BeforeEach
    public void setup() {
        request = new MockHttpServletRequest();
        response = new MockHttpServletResponse();
    }
    
    @Test
    public void cookieSkalKunneHentesUtDekodetFraRequest() {
        
        //Arrange
        request.setCookies(new Cookie[]{new Cookie("User", "Atle+Patle")});
        //Act
        String cookie = CookieUtil.getCookieFromRequest(request, "User");
        //Assert
        assertEquals("Atle Patle", cookie);
    }
    
    @Test
    public void cookieBlirLagretKodetOgTidsmerket() {
        
        //Act
    	CookieUtil.addCookieToResponse(response, "User", "Atle Patle");
        //Assert
        Cookie cookie = response.getCookies()[0];
        assertEquals("User", cookie.getName());
        assertEquals("Atle+Patle", cookie.getValue());
        assertEquals(30, cookie.getMaxAge());
    }

}
